readerMapper + service with active-borrows checking + tests + readerController

+++ src/main/java/com/rest/restlibrary/service/ReaderService.java
@@ -1,34 +1,75 @@
 package com.rest.restlibrary.service;
 
+import com.rest.restlibrary.data.Borrow;
 import com.rest.restlibrary.data.Reader;
+import com.rest.restlibrary.data.dao.BorrowDao;
 import com.rest.restlibrary.data.dao.ReaderDao;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
+import java.lang.management.OperatingSystemMXBean;
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ReaderService {
     @Autowired
     ReaderDao readerDao;
 
-    public void createReader(Reader reader){
+    @Autowired
+    BorrowDao borrowDao;
+
+    public void createReader(Reader reader) {
         readerDao.save(reader);
     }
 
-    public Reader getReader(long readerId){
+    public Reader getReader(long readerId) {
         return readerDao.findOne(readerId);
     }
 
-    public List<Reader> getReaders(){
+    public List<Reader> getReaders() {
         return readerDao.findAll();
     }
 
-    public void updateReader(Reader reader){
+    public void updateReader(Reader reader) {
         readerDao.save(reader);
     }
 
-    public void deleteReader(long readerId){
-        readerDao.delete(readerId);
+    public void deleteReader(long readerId) {
+        boolean hasReaderActiveBorrows = checkReaderHasUnfinishedBorrows(readerId);
+        if (hasReaderActiveBorrows) {
+            throw new RuntimeException("This reader has active borrows, you can not delete");
+        } else {
+            List<Borrow> borrows = readerDao.findOne(readerId).getBorrows();
+            for (Borrow borrow : borrows) {
+                long borrowId = borrow.getId();
+                borrowDao.delete(borrowId);
+            }
+            readerDao.delete(readerId);
+        }
+
+    }
+
+    private boolean checkReaderHasUnfinishedBorrows(long readerId) {
+        List<Borrow> activeBorrows = new ArrayList();
+        List<Borrow> borrows = readerDao.findOne(readerId).getBorrows();
+        if (borrows.isEmpty()) {
+            return false;
+        } else {
+            for (Borrow borrow : borrows) {
+                Optional<LocalDate> untilDate = Optional.ofNullable(borrow.getUntilDate());
+                if (untilDate.isPresent()) {
+                    activeBorrows.add(borrow);
+                }
+            }
+        }
+        if (activeBorrows.isEmpty()) {
+            return false;
+        } else {
+            return true;
+        }
     }
 }
